Introduce a RollbarReporter class
